Split `secp256k1_ec_pubkey_decompress` into in-place and copy variants
Right now `secp256k1_ec_pubkey_decompress` takes an in/out pointer to
a public key and replaces the input key with its decompressed variant.
This forces users who store compressed keys in small (<65 byte) fixed
size buffers (for example, the Rust bindings do this) to explicitly
and wastefully copy their key to a larger buffer.

Add a variant `secp256k1_ec_pubkey_decompress_copy` which takes an
in-pointer and an out-pointer for the public key.